Fix popularity type errors across admin pages (#311)
* Fix TypeError in DataQualityPage when popularity is string

PostgreSQL can return numeric types as strings, causing `.toFixed()`
to fail with "n.toFixed is not a function" error. The page would
render blank due to unhandled runtime error.

Changed `actor.popularity?.toFixed(1)` to
`actor.popularity != null ? Number(actor.popularity).toFixed(1) : "N/A"`
to handle both null values and string-typed numbers.



* Cast popularity to float in all admin API queries

PostgreSQL's numeric(10,3) type is returned as a string by node-pg,
causing frontend .toFixed() calls to fail. Rather than adding
Number() wrappers on every frontend usage, fix at the source by
casting popularity::float in SQL queries.

This fixes the ActorManagementPage crash when filtering by popularity,
and preemptively fixes the same issue in:
- ActorDiagnosticPage
- HighPriorityActorsPage
- CacheManagementPage
- Various enrichment pages



* Address Copilot review feedback

- Simplify DataQualityPage popularity formatting (backend now returns numbers)
- Refactor cache.ts to use parameterized query instead of string interpolation
- Add regression tests for popularity::float cast in SQL queries



---------

+++ server/src/routes/admin/coverage.ts
@@ -182,7 +182,7 @@ router.get("/actors/by-ids", async (req: Request, res: Response): Promise<void> 
       popularity: number | null
       tmdb_id: number | null
     }>(
-      `SELECT id, name, popularity, tmdb_id
+      `SELECT id, name, popularity::float, tmdb_id
        FROM actors
        WHERE id = ANY($1::int[])
        ORDER BY popularity DESC NULLS LAST`,
